gsmarena.py unstable. Trying to complete the parsing of phones
The parse function defines with which callback begin. I did this
this way in order to jump over some steps. Right now I'm jumping
over the root page and the brand page parsing. So the spider
begins directly at parsing the phone page.

Of curse, this implies change the start_url config.